Fix '--' text issue in `heatmap`
I'm not sure why this text is getting added to the ax but this is a simple fix that seems to work on my machine.

In my case with my data (not the collisions dataset), the '--' text only occurred once in the ax. I'm not sure if that's the case for all data, but I leave it up to the maintainers to determine whether or not that is a valid fix for most cases.